move rules extraction to a separate hook
